% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocusPlot.R
\name{LocusPlot}
\alias{LocusPlot}
\title{LocusPlot : produces a pdf file including a plot of region-level results for a set of contiguous regions}
\usage{
LocusPlot(
  chr,
  regionlist,
  pheno = pheno,
  regscanout,
  p_region = 5.86e-07,
  p_single = 5e-08,
  real_size = TRUE,
  allSNPs = FALSE,
  region_tests = NULL,
  outname = NULL
)
}
\description{
LocusPlot : produces a pdf file including a plot of region-level results for a set of contiguous regions
}
