% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiamiPlot.R
\name{MiamiPlot}
\alias{MiamiPlot}
\title{MiamiPlot : produces a pdf file including a miami plot for a pair of tests specified}
\usage{
MiamiPlot(
  regscanout,
  p_region = 5.86e-07,
  p_single = 5e-08,
  allSNPs = FALSE,
  test1 = NULL,
  test2 = NULL,
  outname = NULL,
  chr = NULL,
  zoom = zoom,
  test1_highlight = NULL,
  test2_highlight = NULL
)
}
\description{
MiamiPlot : produces a pdf file including a miami plot for a pair of tests specified
}
