% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLCbinsnpPlot.R
\name{MLCbinsnpPlot}
\alias{MLCbinsnpPlot}
\title{MLCbinsnpPlot : produces a pdf file including the position of the SNPs along the chr per LD bin for the region specified}
\usage{
MLCbinsnpPlot(rscanout, chr_, region_)
}
\description{
MLCbinsnpPlot : produces a pdf file including the position of the SNPs along the chr per LD bin for the region specified
}
