% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqregscanPlot.R
\name{qqregscanPlot}
\alias{qqregscanPlot}
\title{qqregscanPlot : qq plot for a specified test}
\usage{
qqregscanPlot(regionout, outname, test_name)
}
\description{
qqregscanPlot : qq plot for a specified test
}
