% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeVCF.R
\name{recodeVCF}
\alias{recodeVCF}
\title{recodeVCF : prepare input file for region-level analysis}
\usage{
recodeVCF(
  vcfname,
  multiallelic,
  multial_nmaxalleles,
  qc_machr2,
  info_score,
  qcfiler,
  chr = chr,
  start = start,
  end = end
)
}
\description{
recodeVCF : prepare input file for region-level analysis
}
