% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionHeatmap.R
\name{RegionHeatmap}
\alias{RegionHeatmap}
\title{RegionHeatmap : produces heatmaps of correlation matrices within  each region (internal function)}
\usage{
RegionHeatmap(
  inputmat,
  binvector = NULL,
  type_mat = type_mat,
  ptitle,
  colvect = colvect,
  clustorder = clustorder,
  lim = c(-1, 1)
)
}
\description{
RegionHeatmap : produces heatmaps of correlation matrices within  each region (internal function)
}
