% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regscan.R
\name{regscan}
\alias{regscan}
\title{regscan : main function for genome-wide region-level analysis}
\usage{
regscan(
  phenocov = NULL,
  pheno,
  REGIONinfo,
  geno_type,
  pheno_type,
  covlist = NULL,
  data = NULL,
  SNPinfo = NULL,
  vcfname = NULL,
  machr2 = NULL,
  qcinput = NULL,
  info_score = NULL,
  multiallelic = FALSE,
  multial_nmaxalleles = 2,
  mafcut = 0.05,
  rcut = 0.99,
  firthreg = FALSE,
  MLCheatmap = FALSE,
  regionlist = NULL,
  alltests = FALSE,
  edgecut = 0.5,
  tol = 1e-16,
  qcmachr2 = NULL,
  covout = FALSE,
  LDpruning = TRUE,
  singleSNPall = FALSE,
  verbose = FALSE,
  debug = FALSE,
  parallel = FALSE
)
}
\description{
regscan : main function for genome-wide region-level analysis
}
